fix(backend): 6 production bugs found in dry-run feature checks
- schemas/student.py: StudentResponse.middle_name str|None (DB NULL caused 500)
- grace_period_enforcer.py:29: naive vs aware datetime normalize (scheduler TypeError)
- license_crypto.py:708,735: same naive/aware fix (NFC/QR license-gated 500)
- corporate_service.py: create_employee duplicate pre-check raises clean 409
- academic_service.py: create_semester now sets school_id (report-card 404)
- utils/grading.py: float() coercion of avg/max_avg (Decimal += TypeError)

Plus APU Phase 2/3 backend: teacher endpoints, attendance bulk + attendance_batch
model/migration, notifications AUTHENTICATED reads, report-card ownership gate,
FCM push channel (push_devices + fcm_relay), tests (phase3, fcm, teacher, dryrun).

+++ backend/app/utils/grading.py
@@ -13,8 +13,8 @@ def compute_subject_grades(result_map: dict, subject_map: dict) -> tuple[list[di
     total_pct = 0.0
     count = 0
     for subj_id, scores in result_map.items():
-        avg = sum(s["score"] for s in scores) / len(scores)
-        max_avg = sum(s["max_score"] for s in scores) / len(scores)
+        avg = float(sum(s["score"] for s in scores) / len(scores))
+        max_avg = float(sum(s["max_score"] for s in scores) / len(scores))
         pct = round((avg / max_avg) * 100, 1) if max_avg > 0 else 0
         letter = compute_grade(pct)
         subject_grades.append({
